Only consider the first character as the diff symbol.
Previously, the first non-space character of each line was interpreted
as a diff symbol (one of '>', '+', etc.). This is probably too lax,
and can create ambiguous interpretations when the content of the
diffed file itself begins with one of those characters.
Only the first character of each line should be checked for a diff
symbol. Anything else (like a space or any other character) should
signal the whole line as plain text.